+++ contracts/agency/libraries/TickMath.sol
@@ -23,10 +23,10 @@ library TickMath {
   /// @return sqrtPriceX96 A Fixed point Q64.96 number representing the sqrt of the ratio of the two assets (token1/token0)
   ///         at the given tick
   function getSqrtRatioAtTick(int24 tick) internal pure returns (uint160 sqrtPriceX96) {
-    uint256 absTick = tick < 0 ? uint256(-int256(tick)) : uint256(int256(tick));
-    require(absTick <= uint256(int256(MAX_TICK)), 'T');
+    uint absTick = tick < 0 ? uint(-int256(tick)) : uint(int256(tick));
+    require(absTick <= uint(int256(MAX_TICK)), 'T');
 
-    uint256 ratio = absTick & 0x1 != 0 ? 0xfffcb933bd6fad37aa2d162d1a594001 : 0x100000000000000000000000000000000;
+    uint ratio = absTick & 0x1 != 0 ? 0xfffcb933bd6fad37aa2d162d1a594001 : 0x100000000000000000000000000000000;
     if (absTick & 0x2 != 0) ratio = (ratio * 0xfff97272373d413259a46990580e213a) >> 128;
     if (absTick & 0x4 != 0) ratio = (ratio * 0xfff2e50f5f656932ef12357cf3c7fdcc) >> 128;
     if (absTick & 0x8 != 0) ratio = (ratio * 0xffe5caca7e10e4e61c3624eaa0941cd0) >> 128;
@@ -47,7 +47,7 @@ library TickMath {
     if (absTick & 0x40000 != 0) ratio = (ratio * 0x2216e584f5fa1ea926041bedfe98) >> 128;
     if (absTick & 0x80000 != 0) ratio = (ratio * 0x48a170391f7dc42444e8fa2) >> 128;
 
-    if (tick > 0) ratio = type(uint256).max / ratio;
+    if (tick > 0) ratio = type(uint).max / ratio;
 
     // Divides by 1<<32 rounding up to go from a Q128.128 to a Q128.96.
     // we then downcast because we know the result always fits within 160 bits due to our tick input constraint
@@ -62,10 +62,10 @@ library TickMath {
   function getTickAtSqrtRatio(uint160 sqrtPriceX96) internal pure returns (int24 tick) {
     // Second inequality must be < because the price can never reach the price at the max tick
     require(sqrtPriceX96 >= MIN_SQRT_RATIO && sqrtPriceX96 < MAX_SQRT_RATIO, 'R');
-    uint256 ratio = uint256(sqrtPriceX96) << 32;
+    uint ratio = uint(sqrtPriceX96) << 32;
 
-    uint256 r = ratio;
-    uint256 msb = 0;
+    uint r = ratio;
+    uint msb = 0;
 
     assembly {
       let f := shl(7, gt(r, 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF))
