+++ contracts/core/interfaces/ICallee.sol
@@ -4,8 +4,8 @@ pragma solidity ^0.8.0;
 interface ICallee {
     function call(
         address sender,
-        uint256 amount0,
-        uint256 amount1,
+        uint amount0,
+        uint amount1,
         bytes calldata data
     ) external;
 }
