+++ contracts/for-test/ERC20ForTest.sol
@@ -8,23 +8,23 @@ contract ERC20ForTest is ERC20 {
     string memory _name,
     string memory _symbol,
     address _initialAccount,
-    uint256 _initialBalance
+    uint _initialBalance
   ) ERC20(_name, _symbol) {
     _mint(_initialAccount, _initialBalance);
   }
 
-  function mint(address _account, uint256 _amount) public {
+  function mint(address _account, uint _amount) public {
     _mint(_account, _amount);
   }
 
-  function burn(address _account, uint256 _amount) public {
+  function burn(address _account, uint _amount) public {
     _burn(_account, _amount);
   }
 
   function transferInternal(
     address _from,
     address _to,
-    uint256 _value
+    uint _value
   ) public {
     _transfer(_from, _to, _value);
   }
@@ -32,12 +32,12 @@ contract ERC20ForTest is ERC20 {
   function approveInternal(
     address _owner,
     address _spender,
-    uint256 _value
+    uint _value
   ) public {
     _approve(_owner, _spender, _value);
   }
 
-  function deposit(uint256 _amount) external payable {
+  function deposit(uint _amount) external payable {
     // Function added for compatibility with WETH
   }
 }
