+++ contracts/for-test/JobForTest.sol
@@ -6,7 +6,7 @@ import '../interfaces/IKeep3r.sol';
 contract JobForTest {
   error InvalidKeeper();
   address public keep3r;
-  uint256 public nonce;
+  uint public nonce;
 
   constructor(address _keep3r) {
     keep3r = _keep3r;
@@ -15,17 +15,17 @@ contract JobForTest {
   function work() external {
     if (!IKeep3r(keep3r).isKeeper(msg.sender)) revert InvalidKeeper();
 
-    for (uint256 i; i < 1000; i++) {
+    for (uint i; i < 1000; i++) {
       nonce++;
     }
 
     IKeep3r(keep3r).worked(msg.sender);
   }
 
-  function workHard(uint256 _factor) external {
+  function workHard(uint _factor) external {
     if (!IKeep3r(keep3r).isKeeper(msg.sender)) revert InvalidKeeper();
 
-    for (uint256 i; i < 1000 * _factor; i++) {
+    for (uint i; i < 1000 * _factor; i++) {
       nonce++;
     }
