+++ contracts/interfaces/external/IWeth9.sol
@@ -6,5 +6,5 @@ import '@openzeppelin/contracts/token/ERC20/IERC20.sol';
 interface IWeth9 is IERC20 {
   function deposit() external payable;
 
-  function withdraw(uint256) external;
+  function withdraw(uint) external;
 }
