+++ contracts/mock/MockToken.sol
@@ -7,11 +7,11 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 contract MockToken is ERC20, Ownable {
     constructor(string memory _name, string memory _symbol) ERC20(_name, _symbol) {}
 
-    function mint(address _to, uint256 _amount) external onlyOwner {
+    function mint(address _to, uint _amount) external onlyOwner {
         _mint(_to, _amount);
     }
 
-    function burn(address _to, uint256 _amount) external onlyOwner {
+    function burn(address _to, uint _amount) external onlyOwner {
         _burn(_to, _amount);
     }
 }
